Time: 0 ms (100.00%), Space: 6.2 MB (71.47%) - LeetHub

+++ 0367-valid-perfect-square/0367-valid-perfect-square.cpp
@@ -6,10 +6,10 @@ public:
         int start = 1;
         int end = num;
         while(start <= end){
-            long long mid = start + (end-start)/2;
-            if(mid*mid > (long long) num)
+            double mid = start + (end-start)/2;
+            if(mid*mid >  num)
                 end = mid - 1.0;
-            else if(mid*mid == (long long) num)
+            else if(mid*mid ==  num)
                 return true;
             else
                 start = mid + 1.0;
